+++ mapsforge-map-awt/src/test/java/org/mapsforge/map/layer/overlay/PolylineTest.java
@@ -1,6 +1,7 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014 Ludwig M Brinckmann
+ * Copyright 2018 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -26,7 +27,9 @@ import org.mapsforge.core.model.Point;
 import org.mapsforge.map.awt.graphics.AwtGraphicFactory;
 import org.mapsforge.map.model.FixedTileSizeDisplayModel;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 public class PolylineTest {
     private static final GraphicFactory GRAPHIC_FACTORY = AwtGraphicFactory.INSTANCE;
@@ -53,8 +56,10 @@ public class PolylineTest {
             Point point = new Point(0, 0);
             polyline.draw(boundingBox, (byte) 0, canvas, point);
 
-            polyline.getLatLongs().add(new LatLong(0, 0));
-            polyline.getLatLongs().add(new LatLong(1, 1));
+            List<LatLong> latLongs = new ArrayList<>();
+            latLongs.add(new LatLong(0, 0));
+            latLongs.add(new LatLong(1, 1));
+            polyline.setPoints(latLongs);
             polyline.draw(boundingBox, (byte) 0, canvas, point);
 
             polyline.setPaintStroke(GRAPHIC_FACTORY.createPaint());
@@ -71,7 +76,9 @@ public class PolylineTest {
         Assert.assertTrue(polyline.getLatLongs().isEmpty());
         Assert.assertNull(polyline.getPaintStroke());
 
-        polyline.getLatLongs().add(latLong);
+        List<LatLong> latLongs = new ArrayList<>();
+        latLongs.add(latLong);
+        polyline.setPoints(latLongs);
         Assert.assertEquals(Arrays.asList(latLong), polyline.getLatLongs());
 
         polyline.setPaintStroke(paintStroke);